+++ src/pkg_about/_about.py
@@ -46,7 +46,7 @@ def about(package=None):
         __maintainer__       = metadata.get("Maintainer"),
         __maintainer_email__ = metadata.get("Maintainer-email"),
         __license__      = metadata.get("License"),
-        __copyright__    = __get_copyright(metadata.get("Description"))
+        __copyright__    = metadata.get("Author"),
     )
 
     pkg_globals.update(pkg_metadata)
@@ -55,7 +55,6 @@ def about(package=None):
 
 def about_from_setup(package_path=None):
     import sys
-    import re
     from pathlib import Path
     from packaging.version import parse as parse_version
     try:
@@ -73,7 +72,6 @@ def about_from_setup(package_path=None):
                     if package_path is None else Path(package_path))
     pyproject_path = package_path/"pyproject.toml"
     setup_cfg_path = package_path/"setup.cfg"
-    readme_path    = package_path/"README.rst"
     metadata = {}
     if setup_cfg_path.exists():  # pragma: no branch
         metadata.update(read_setupcfg(setup_cfg_path,
@@ -89,11 +87,8 @@ def about_from_setup(package_path=None):
                 import tomli as tomllib
             with pyproject_path.open("rb") as file:
                 metadata.update(tomllib.load(file).get("project", {}))
-    copyr_patt = re.compile(r"^\s*__copyright__\s*=\s*")
-    about_py = package_path.glob("src/**/__about__.py")
     version = parse_version(metadata["version"])
-    release_levels = __release_levels
-    get, get_copyright = __get, __get_copyright
+    get, release_levels = __get, __release_levels
 
     class about:
         __slots__  = ()
@@ -128,14 +123,10 @@ def about_from_setup(package_path=None):
                                 or get(metadata, "maintainer"))
         __maintainer_email__ = (get(metadata, "maintainers", 1, "email")
                                 or get(metadata, "maintainer_email"))
-        __license__   = (get(metadata, "license", "text")
-                         or get(metadata, "license"))
-        __copyright__ = eval(next((copyr_patt.split(line)[1] for line in
-                                   (next(about_py).open("rt", encoding="utf-8")
-                                    if about_py else ())
-                                   if copyr_patt.split(line)[1:]), "None"))
-        if __copyright__ is None and readme_path.exists():  # pragma: no branch
-            __copyright__ = get_copyright(readme_path.read_text(encoding="utf-8"))
+        __license__      = (get(metadata, "license", "text")
+                            or get(metadata, "license"))
+        __copyright__    = (get(metadata, "authors", 0, "name")
+                            or get(metadata, "author"))
 
     pkg_globals["about"] = about
     pkg_globals.setdefault("__all__", [])
@@ -154,28 +145,6 @@ def __get(mdata, *keys):
             return None
         mdata = mdata[key]
     return mdata
-
-
-def __get_copyright(description):
-    from docutils.core import publish_doctree
-    from docutils import nodes
-    copyr = None
-    if description is not None:  # pragma: no branch
-        document = publish_doctree(description)
-        subst_name = document.substitution_names.get("copyright")
-        substitution = document.substitution_defs.get(subst_name)
-        if substitution is not None:
-            copyr = substitution.astext()
-        else:
-            # Try to get from 'License' section
-            section = document.ids.get(document.nameids.get("license"))
-            if section is not None:  # pragma: no branch
-                lblock = section.next_node(nodes.line_block)
-                if lblock is not None:  # pragma: no branch
-                    copyr = next((line for _ in lblock.findall(nodes.line)
-                                  if ((line := _.astext().lstrip()).lower()
-                                      .startswith("copyright"))), None)
-    return copyr
 
 
 __release_levels = dict(